requirements.txt
numpy==1.15.4
Pillow==5.3.0
six==1.11.0
torch-nightly==1.0.0.dev20181113
torchvision-nightly==0.2.1
tqdm==4.28.1
